requirements.txt

viam-sdk==0.58.0

typing-extensions
python-vlc
Pillow